confirmation on delete video || removed flip camra option

+++ Dashboard/index.php
@@ -48,7 +48,7 @@ mysqli_close($conn);
                     <a href="single.php?vid=<?php echo $video['video'] ?>" class="video-goto">
                       <ion-icon name="play-outline"></ion-icon>
                     </a>
-                    <a href="deleteVid.php?vid=<?php echo $video['video'] ?>" class="delete-bt">
+                    <a href="deleteVid.php?vid=<?php echo $video['video'] ?>" onclick="return confirm('Are you sure you want to delete a video?')" class="delete-bt">
                       <ion-icon name="trash-outline"></ion-icon>
                     </a>
                     <div class="video-title">
